Refactor static and common functions to utility module
